Demote Symtab.scopify to private Outliner.scopify
